+++ packages/ui-table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @buildpad/ui-table
 
+## 1.10.0
+
+### Patch Changes
+
+- Updated dependencies [5981327]
+  - @buildpad/types@1.10.0
+  - @buildpad/services@1.10.0
+  - @buildpad/hooks@1.10.0
+
 ## 1.9.3
 
 ### Patch Changes